+++ src/decorator/columns/UpdateDateColumn.ts
@@ -9,21 +9,11 @@ import {ColumnMetadataArgs} from "../../metadata-args/ColumnMetadataArgs";
 export function UpdateDateColumn(options?: ColumnOptions): Function {
     return function (object: Object, propertyName: string) {
 
-        // const reflectedType = ColumnTypes.typeToString((Reflect as any).getMetadata("design:type", object, propertyName));
-
-        // if column options are not given then create a new empty options
-        if (!options) options = {} as ColumnOptions;
-
-        // implicitly set a type, because this column's type cannot be anything else except date
-        // options = Object.assign({ type: Date } as ColumnOptions, options);
-
-        // create and register a new column metadata
         const args: ColumnMetadataArgs = {
             target: object.constructor,
             propertyName: propertyName,
-            // propertyType: reflectedType,
             mode: "updateDate",
-            options: options
+            options: options ? options : {}
         };
         getMetadataArgsStorage().columns.push(args);
     };