+++ src/driver/mysql/MysqlDriver.ts
@@ -86,11 +86,13 @@ export class MysqlDriver implements Driver {
      */
     mappedDataTypes: MappedColumnTypes = {
         createDate: "datetime",
+        createDateDefault: "CURRENT_TIMESTAMP",
         updateDate: "datetime",
+        updateDateDefault: "CURRENT_TIMESTAMP",
         version: "int",
         treeLevel: "int",
         migrationName: "varchar",
-        migrationTimestamp: "timestamp",
+        migrationTimestamp: "timestamp"
     };
 
     // -------------------------------------------------------------------------
@@ -284,6 +286,10 @@ export class MysqlDriver implements Driver {
             type += column.type;
         }
 
+        // normalize shortcuts
+        if (type === "integer")
+            type = "int";
+
         if (column.length) {
             type += "(" + column.length + ")";
 
@@ -320,6 +326,27 @@ export class MysqlDriver implements Driver {
             type += "(4)";
 
         return type;
+    }
+
+    /**
+     * Normalizes "default" value of the column.
+     */
+    normalizeDefault(column: ColumnMetadata): string {
+        if (typeof column.default === "number") {
+            return "" + column.default;
+
+        } else if (typeof column.default === "boolean") {
+            return column.default === true ? "1" : "0";
+
+        } else if (typeof column.default === "function") {
+            return column.default();
+
+        } else if (typeof column.default === "string") {
+            return `'${column.default}'`;
+
+        } else {
+            return column.default;
+        }
     }
 
     // -------------------------------------------------------------------------